chore(ffi): move RoomInfo to the `room` module
