+++ bindings/matrix-sdk-ffi/src/room/room_info.rs
